build(deps): update cocache version to 3.0.0 - Update cocache dependency from 2.6.3 to 3.0.0 - Refactor import statements to use api packages instead of internal implementations - Update CacheValue and TtlAt usages to their new package locations
