+++ cosec-redis/src/test/kotlin/me/ahoo/cosec/redis/PolicyCodecExecutorTest.kt
@@ -12,8 +12,9 @@
  */
 package me.ahoo.cosec.redis
 
-import me.ahoo.cache.CacheValue
-import me.ahoo.cache.TtlAt
+import me.ahoo.cache.ComputedTtlAt
+import me.ahoo.cache.DefaultCacheValue
+import me.ahoo.cache.api.CacheValue
 import me.ahoo.cache.spring.redis.codec.ObjectToJsonCodecExecutor
 import me.ahoo.cosec.api.policy.Policy
 import me.ahoo.cosec.api.policy.PolicyType
@@ -67,18 +68,18 @@ internal class PolicyCodecExecutorTest {
             statements = listOf(StatementData(action = AllActionMatcher.INSTANCE)),
         )
         val key = "policy:" + MockIdGenerator.INSTANCE.generateAsString()
-        val cacheValue: CacheValue<Policy> = CacheValue.forever(policy)
+        val cacheValue: CacheValue<Policy> = DefaultCacheValue.forever(policy)
         codecExecutor.executeAndEncode(key, cacheValue)
-        val (value) = codecExecutor.executeAndDecode(key, TtlAt.FOREVER)
+        val value = codecExecutor.executeAndDecode(key, ComputedTtlAt.FOREVER).value
         assertThat(value, `is`(policy))
     }
 
     @Test
     fun executeAndEncodeMissing() {
         val key = "policy:" + MockIdGenerator.INSTANCE.generateAsString()
-        val cacheValue: CacheValue<Policy> = CacheValue.missingGuard()
+        val cacheValue: CacheValue<Policy> = DefaultCacheValue.missingGuard()
         codecExecutor.executeAndEncode(key, cacheValue)
-        val actual = codecExecutor.executeAndDecode(key, TtlAt.FOREVER)
+        val actual = codecExecutor.executeAndDecode(key, ComputedTtlAt.FOREVER)
         assertThat(actual, `is`(cacheValue))
     }
 }
