+++ gradle/libs.versions.toml
@@ -3,7 +3,7 @@
 spring-boot = "3.4.3"
 spring-cloud = "2024.0.0"
 cosid = "2.11.0"
-cocache = "2.6.3"
+cocache = "3.0.0"
 opentelemetry = "2.13.2"
 testcontainers = "1.20.5"
 guava = "33.4.0-jre"
